Create the neet.states generator
The new generator generates lazily generates all possible states
of a network given a specification, `spec`. The specification may
either be an integer or a list of integers.

If `spec` is an integer, the network is assumed to be boolean with
as many nodes as `spec`. On the other hand, if `spec` is a list,
then the network is assumed to have `len(spec)` nodes and the
elements of `spec` are assumed to be the bases of those nodes.