Fix range error under Python 3
We were using xrange to ensure that a loop was over a generator
as opposed to a list. Of course xrange does not exist in Python 3,
so we've switch to range.

This isn't a real problem in this case since if, under Python 2,
the use of range gives you memory problems, then you probably have
bigger memory issues to worry about.