Handle base mismatch
StateSpace objects can be constructed from a list of bases or from
a number of nodes and a base. It is possible to provide a list of
bases and a base, which is somewhat nonsensical. We now check to
ensure that if a base is provided along with a list of bases, that
the list of bases is uniform and that the provided base is the
same. In any other case, we raise an error.