Reduce safe calls in asynchronous.transitions
The transitions function was calling the update and encode methods
which check the validity of their arguments. This is unnecessary
as other mechanisms ensure that the arguments are valid. We've
replace (most) of those calls with calls to unsafe alternatives.

In addition, we only encode states if the encoding is necessary.
Otherwise, tuples are used for dictionary keys.