Replace trajectory with transitions timeseries
The timeseries function was calling trajectory repeatedly which
would subsequently call encode and update. This became very costly
as most states in the network have multiple preimages and thus
were being computed and encoded excessively.

To remedy this, we make a single call to transitions and use the
resulting array as a lookup table. This is then used to populate
the time series.