Roll a manual implementation of attractors
The networkx-based algorithm for computing the attractors of a
network was excessively memory intensive. This was preventing us
from analyzing networks with more than about 20 nodes. This new
algorithm uses substantially less memory at the expense of some
runtime performance. We will be investing effort into improving
the speed in a later commit.