Create the `Landscape.trajectory` method
This method takes a state and computes the trajectory of that state
within the landscape. The trajectory stops when once a state in
the trajectory has been seen previously, i.e. when the entire
attractor cycle of the state has been traversed, unless a specified
number of time steps is requested.