Create the `Landscape.heights` property
This property returns an array of the heights of each state. We
define the height of each state as the number of steps from the
state to an attractor state. By this definition, the height of a
state in an attractor is 0.

The recurrence time of a state is the height of the state plus the
length of it's associated attractor cycle. We'll be adding this
property in the next couple of commits.