Encode states as long not int
When we move to large networks, e.g. greater than 63 nodes, we
end up with integer overflows which numpy forces to floating-point
numbers instead of long. Since states need to be integers, we make
sure that all states are encoded as longs so as to never overflow.

This has some performance cost, but allows us to complete
computations, e.g. sensitivity, for large networks where we would
otherwise have had overflow errors.